test: 노선 단위 테스트 추가 및 중복 테스트 제거
